// src/types/property.type.ts
export interface FilteredPropertyParams {
  kinds?: ("SHARE" | "RENT")[];
  sharePropertySubTypes?: string[];
  rentPropertySubTypes?: string[];
  minWeeklyCost?: number;
  maxWeeklyCost?: number;
  billIncluded?: boolean;
  availableFrom?: string;
  availableTo?: string;
  immediate?: boolean;
  negotiable?: boolean;
  metroStopLatitude?: number;
  metroStopLongitude?: number;
  radiusKm?: number;
  suburb: string;
}